renamed StackAbstract to AbstractStack
StackAbstract still here for BC reasons
